+++ src/main/java/org/openlmis/referencedata/util/messagekeys/LotMessageKeys.java
@@ -25,6 +25,8 @@ public abstract class LotMessageKeys extends MessageKeys {
   public static final String ERROR_TRADE_ITEM_REQUIRED = join(ERROR, TRADE_ITEM, REQUIRED);
   public static final String ERROR_NOT_FOUND = join(ERROR, NOT_FOUND);
   public static final String ERROR_LOT_CODE_MUST_BE_UNIQUE = join(ERROR, LOT_CODE, MUST_BE_UNIQUE);
+  public static final String ERROR_LOT_CODE_TOO_LONG = join(ERROR, LOT_CODE, "tooLong");
+  public static final String ERROR_LOT_CODE_INVALID_FORMAT = join(ERROR, LOT_CODE, "invalidFormat");
   public static final String ERROR_ORDERABLE_ID_AND_TRADE_ITEM_ID_USED_TOGETHER =
           join(ERROR, NOT_TOGETHER, "orderableIdAndtradeItemdId");
 }